requirements.txt
google-genai>=0.8.0
python-dotenv>=1.0.0
pytest>=8.0.0
